refactor: move helper to unit tests
